gq develop US2025091002 split analysis config window
