+++ ui/custom_ui_widget/custom_plot_layout_widget.py
@@ -1,0 +1,41 @@
+import pyqtgraph as pg
+from PyQt5.QtGui import QIcon, QFont
+from PyQt5.QtWidgets import QVBoxLayout, QWidget
+
+from consts.running_consts import DEFAULT_DIR
+from ui.graph_widget import custom_log_tick_strings
+
+class AnalysisGraphWidget(QWidget):
+
+    def __init__(self):
+        super().__init__()
+
+        self.analysis_plot = pg.PlotWidget()
+
+        self.set_plot_font_size(20)
+        self.init_ui()
+
+    def init_ui(self):
+        self.setWindowIcon(QIcon(DEFAULT_DIR + "ui/ui_pic/logo_pic/ting.ico"))
+
+        self.analysis_plot.setBackground("white")
+
+        layout = QVBoxLayout()
+        layout.addWidget(self.analysis_plot)
+        self.setLayout(layout)
+
+    def set_plot_font_size(self, font_size: int):
+        font = QFont()
+        font.setPixelSize(font_size)
+
+        b_axis = self.analysis_plot.getAxis("bottom")
+        l_axis = self.analysis_plot.getAxis("left")
+
+        b_axis.logTickStrings = custom_log_tick_strings
+
+        b_axis.setTickFont(font)
+        l_axis.setTickFont(font)
+        b_axis.setTextPen("black")
+        l_axis.setTextPen("black")
+        b_axis.setLabel(b_axis.labelText, **{"font-size": f"{font_size}px"})
+        l_axis.setLabel(l_axis.labelText, **{"font-size": f"{font_size}px"})
